Pin the netty-transport-native-epoll to avoid conflicts (#6194)
### Motivation

Currently the version pinning for `netty-transport-native-epoll` is not including the native library artifact.

That results, depending on the Maven version, to be picking up an earlier version of `transport-native-epoll-4.1.33.Final-linux-x86_64.jar`, where the version is 4.1.33 as opposed to 4.1.43 which is the correct expected version.

This results in using Java NIO based transport instead of the more effiecient/performant epoll based one.

This affects 2.5.0 as well.
(cherry picked from commit 857d63b84d0fcf50ca93d7d5020ec88f1255933d)