client: make SubscriptionMode a member of ConsumerConfigurationData (#6337)
Currently, SubscriptionMode is a parameter to create ConsumerImpl, but it is not exported out, and user could not set this value for consumer.  This change tries to make SubscriptionMode a member of ConsumerConfigurationData, so user could set this parameter when create consumer.
(cherry picked from commit 208af7cd7e718d378707df4aba600be3202e23df)