[proxy] Fix proxy routing to functions worker (#6486)
### Motivation

Currently, the proxy only works to proxy v1/v2 functions routes to the
function worker.

### Modifications

This changes this code to proxy all routes for the function worker when
those routes match. At the moment this is still a static list of
prefixes, but in the future it may be possible to have this list of
prefixes be dynamically fetched from the REST routes.

### Verifying this change
- added some tests to ensure the routing works as expected
(cherry picked from commit 329e2310069b61e25ce3f87f2828fab78f97187a)